requirements.txt
cassandra-driver
redis
jsonpickle
discord.py
grpcio-tools
protobuf-compiler
#google apis?
google-cloud-storage
